Add files via upload
Data Dictionary

RunsScored : Number of Runs scored by Kohli in the match.
BallsFaced : Number of Balls Faces by Kohli in the match.
BattingPosition : Batting position is Virat Kohli's position on the batting order.
Dismissal : How Kohli got out in the match or not out. Caught, NO: Not Out, Bowled, Run out, LBW, Stumped, Hit wicket.
Inning : An innings is one of the divisions of a match during which one team takes its turn to bat. Two 1st inning or 2nd inning.
Opposition : Name of the Opposition team.
Year : Year in which the match was played.